// www/js/AccessManagementUpsellV2/ts/react/legallySensitiveContent/enums/ConsentName.ts
export enum ConsentName {
  phoneNumberDiscoverabilitySetting = 'phoneNumberDiscoverabilitySetting',
  phoneNumberDiscoverabilitySettingParentSide = 'phoneNumberDiscoverabilitySettingParentSide',
  phoneNumberDiscoverabilityUpsell = 'phoneNumberDiscoverabilityUpsell',
  personalizedAdsSetting = 'personalizedAdsSetting',
  sellShareDataSetting = 'sellShareDataSetting',
  allowMarketingEmailCheckboxEmailVerification = 'allowMarketingEmailCheckboxEmailVerification',
  voiceDataConsentSetting = 'voiceDataConsentSetting',
  voiceDataConsentSettingParentSide = 'voiceDataConsentSettingParentSide',
  whoCanPartyWithMe = 'whoCanPartyWithMe',
  whoCanPartyWithMeV2 = 'whoCanPartyWithMeV2',
  whoCanPartyWithMeParentSide = 'whoCanPartyWithMeParentSide',
  whoCanPartyWithMeParentSideV2 = 'whoCanPartyWithMeParentSideV2',
  whoCanUsePartyChatWithMe = 'whoCanUsePartyChatWithMe',
  whoCanUsePartyChatWithMeV2 = 'whoCanUsePartyChatWithMeV2',
  whoCanUsePartyChatWithMeParentSide = 'whoCanUsePartyChatWithMeParentSide',
  whoCanUsePartyChatWithMeParentSideV2 = 'whoCanUsePartyChatWithMeParentSideV2',
  whoCanUsePartyVoiceWithMe = 'whoCanUsePartyVoiceWithMe',
  whoCanUsePartyVoiceWithMeV2 = 'whoCanUsePartyVoiceWithMeV2',
  whoCanUsePartyVoiceWithMeParentSide = 'whoCanUsePartyVoiceWithMeParentSide',
  whoCanUsePartyVoiceWithMeParentSideV2 = 'whoCanUsePartyVoiceWithMeParentSideV2',
  allowMarketingEmailNotifications = 'allowMarketingEmailNotifications',
  receiveRobuxTransferConsentCard = 'receiveRobuxTransferConsentCard',
  sendRobuxTransferConsentCard = 'sendRobuxTransferConsentCard',
  vpcRequestLinkSubjectToPC = 'vpcRequestLinkSubjectToPC',
  vpcRequestLinkNotSubjectToPC = 'vpcRequestLinkNotSubjectToPC',
  vpcRequestLinkDefault = 'vpcRequestLinkDefault',
  consentCenterAllowAction = 'consentCenterAllowAction',
  consentCenterUpdateSettingNoValue = 'consentCenterUpdateSettingNoValue',
  consentCenterUpdateSettingWithValue = 'consentCenterUpdateSettingWithValue',
  whoCanPartyWithMeTrustedFriends = 'whoCanPartyWithMeTrustedFriends',
  whoCanPartyWithMeTrustedFriendsV2 = 'whoCanPartyWithMeTrustedFriendsV2',
  whoCanPartyWithMeParentSideRemovedComms = 'whoCanPartyWithMeParentSideRemovedComms',
  whoCanPartyWithMeParentSideRemovedCommsV2 = 'whoCanPartyWithMeParentSideRemovedCommsV2',
  whoCanUsePartyChatWithMeTrustedFriends = 'whoCanUsePartyChatWithMeTrustedFriends',
  whoCanUsePartyChatWithMeTrustedFriendsV2 = 'whoCanUsePartyChatWithMeTrustedFriendsV2',
  whoCanUsePartyChatWithMeParentSideTrustedFriends = 'whoCanUsePartyChatWithMeParentSideTrustedFriends',
  whoCanUsePartyChatWithMeParentSideTrustedFriendsV2 = 'whoCanUsePartyChatWithMeParentSideTrustedFriendsV2',
  whoCanUsePartyVoiceWithMeTrustedFriends = 'whoCanUsePartyVoiceWithMeTrustedFriends',
  whoCanUsePartyVoiceWithMeTrustedFriendsV2 = 'whoCanUsePartyVoiceWithMeTrustedFriendsV2',
  whoCanUsePartyVoiceWithMeParentSideTrustedFriends = 'whoCanUsePartyVoiceWithMeParentSideTrustedFriends',
  whoCanUsePartyVoiceWithMeParentSideTrustedFriendsV2 = 'whoCanUsePartyVoiceWithMeParentSideTrustedFriendsV2'
}

export default ConsentName;